Feature/30 Code Quality fixes

+++ acropolis-service/src/main/kotlin/org/ephyra/acropolis/service/api/IApplicationSoftwareService.kt
@@ -1,5 +1,5 @@
 package org.ephyra.acropolis.service.api
 
 interface IApplicationSoftwareService {
-    fun create(projectId: Int, name: String)
+    fun create(projectId : Long, name: String)
 }
