+++ acropolis-service/src/main/kotlin/org/ephyra/acropolis/service/impl/SystemSoftwareService.kt
@@ -12,7 +12,7 @@ class SystemSoftwareService : ISystemSoftwareService {
     @Autowired
     private lateinit var persistence: SystemSoftwarePersistence
 
-    override fun create(projectId: Int, name: String) {
+    override fun create(projectId : Long, name: String) {
         val project = ProjectEntity()
         project.id = projectId
